+++ examples/test_mtmdd.cc
@@ -48,7 +48,7 @@ using namespace MEDDLY;
 #if USE_REALS
   typedef float element_type;
 #else
-  typedef int element_type;
+  typedef long element_type;
 #endif
 
 
@@ -251,7 +251,7 @@ int main(int argc, char *argv[])
 #if USE_REALS
       printf(": %f]\n", terms[i]);
 #else
-      printf(": %d]\n", terms[i]);
+      printf(": %ld]\n", terms[i]);
 #endif
     }
   }
